Added Multimode compatible PCB
As I was finished with the Genesis PCB I wanted to to the next one and stumbled upon the multimode function.
By connecting one analog pin to ground on the Arduino it can swap to different mode slots.

I changed the DB9 to be basically a 1:1 passthrough 1-8 are D2-D9 and pin 9 is grounded.
It makes it possible to just get DB-9 breakouts with screw terminals and split and splice the cable in there in for all consoles that use a common ground and do not use more than 8 wires for data. 
I added a DIP switch to the analog side of the Arduino to switch between the console modes.